注册功能加入缓存和bloom过滤器 将long id改为int

+++ src/main/java/com/syclover/geekPlatform/dao/AuthorMapper.java
@@ -13,7 +13,7 @@ import java.util.List;
 public interface AuthorMapper {
 
     // 根据id获取出题人
-    Author getById(long id);
+    Author getById(int id);
 
     // 获得所有出题人
     List<Author> getAll();
